fix(from-json-tear-off): adjust type signature of fromJson tear-off
fixes https://github.com/rrousselGit/freezed/issues/517

this change swaps the type signature of the generated fromJson tear-off
to Map<String, Object?> instead of Map<String, Object>.

it seems most likely that this was missed in the upgrade to null-safety.
this essentially restores the previously intended behavior. maybe once
the next version of dart lands we won't need to generate these anymore
:) but for now it's super useful when it works!