Reinstate @With.fromString and @Implements.fromString
These were removed in v 0.15.0 but it turns out they are necessary when
the mixed in or implemented interface is generic. Freezed would generate
correct code but dart would throw the following load error caused by the
annotation declaration when compiling:

Error: Type variables can't be used in static members.