docs: simplify dependency installation commands by grouping packages
Allows users to copy-paste commands and run them at once instead of copying each command and running it separately. This is also a bit more readable since `build_runner` is always required when using `freezed`, and both `freezed` and `freezed_annotation` are required at the same time to use Freezed. While `json_serializable` can be slightly unrelated since serialization is not part of data classes, and it makes sense to separate them in case the user doesn't need JSON serialization.